+++ components/restreamer/client/src/components/ClientStatistics.svelte
@@ -1,5 +1,5 @@
 <script lang="js">
-  import { statusesList } from '../constants/statuses';
+  import { statusesList } from '../utils/constants';
   import Confirm from './common/Confirm.svelte';
   import { showError } from '../utils/util';
   import { mutation } from 'svelte-apollo';