+++ components/restreamer/client/src/utils/constants.js
@@ -4,3 +4,8 @@ export const ONLINE = 'ONLINE';
 export const UNSTABLE = 'UNSTABLE';
 
 export const statusesList = [OFFLINE, INITIALIZING, ONLINE, UNSTABLE];
+
+export const STREAM_ERROR = 'STREAM_ERROR';
+export const STREAM_WARNING = 'STREAM_WARNING';
+
+export const streamStatusList = [STREAM_ERROR, STREAM_WARNING];