+++ components/restreamer/client/src/utils/statusFilters.util.js
@@ -1,7 +1,0 @@
-export const toggleFilterStatus = (filters, filter) => {
-  const filterIndex = filters.indexOf(filter);
-
-  return filterIndex === -1
-    ? [...filters, filter]
-    : filters.filter((item) => item !== filter);
-};