+++ components/restreamer/src/state/client_statistics.rs
@@ -4,6 +4,7 @@
 //! [`Input`]: crate::state::Input
 //! [`Output`]: crate::state::Output
 use crate::state::Status;
+use anyhow::anyhow;
 use chrono::{DateTime, Utc};
 
 use derive_more::{Deref, Display, Into};
@@ -12,6 +13,7 @@ use juniper::{
     ParseScalarValue, ScalarToken, ScalarValue, Value,
 };
 
+use crate::{stream_probe::StreamInfo, types::UNumber};
 use serde::{Deserialize, Deserializer, Serialize};
 use url::Url;
 
@@ -236,5 +238,91 @@ impl<'de> Deserialize<'de> for ClientId {
         D: Deserializer<'de>,
     {
         Ok(Self::new(Url::deserialize(deserializer)?))
+    }
+}
+
+/// Stream statistics
+#[derive(
+    Clone, Debug, Deserialize, Eq, Serialize, PartialEq, GraphQLObject,
+)]
+pub struct StreamStatistics {
+    /// Name of audio codec.  Example: "aac"
+    pub audio_codec_name: Option<String>,
+    /// Stereo / Mono layout
+    pub audio_channel_layout: Option<String>,
+    // Audio sample rate. Example - "44100"
+    pub audio_sample_rate: Option<String>,
+    /// Count of audio channels. Example: 2
+    pub audio_channels: Option<UNumber>,
+    // Name of video codec. Example: "h264"
+    pub video_codec_name: Option<String>,
+    // Video frame rate (fps). Example: "30/1"
+    pub video_r_frame_rate: Option<String>,
+    /// Video width
+    pub video_width: Option<UNumber>,
+    // Video height
+    pub video_height: Option<UNumber>,
+    // Total bit rate
+    pub bit_rate: Option<String>,
+    // Error message, if we could not retrieve stream info
+    pub error: Option<String>,
+}
+
+impl StreamStatistics {
+    /// Constructs [`StreamStatistics`] from [`Result`]
+    #[must_use]
+    pub fn new(result: anyhow::Result<StreamInfo>) -> Self {
+        match result {
+            Err(e) => Self::create_error_instance(&e),
+            Ok(info) => {
+                let audio_stream = match info.find_stream("audio") {
+                    Some(audio) => audio,
+                    None => {
+                        return Self::create_error_instance(&anyhow!(
+                            "Can't find 'audio' stream"
+                        ))
+                    }
+                };
+
+                let video_stream = match info.find_stream("video") {
+                    Some(video) => video,
+                    None => {
+                        return Self::create_error_instance(&anyhow!(
+                            "Can't find 'video' stream"
+                        ))
+                    }
+                };
+
+                Self {
+                    audio_codec_name: audio_stream.codec_name,
+                    audio_channel_layout: audio_stream.channel_layout,
+                    audio_sample_rate: audio_stream.sample_rate,
+                    audio_channels: audio_stream
+                        .channels
+                        .map(|x| UNumber::new(x.into())),
+                    video_codec_name: video_stream.codec_name,
+                    video_r_frame_rate: video_stream.r_frame_rate,
+                    video_width: video_stream.width.map(UNumber::new),
+                    video_height: video_stream.height.map(UNumber::new),
+                    bit_rate: info.format.bit_rate,
+                    error: None,
+                }
+            }
+        }
+    }
+
+    pub fn create_error_instance(e: &anyhow::Error) -> Self {
+        Self {
+            audio_codec_name: None,
+            audio_channel_layout: None,
+            audio_sample_rate: None,
+            audio_channels: None,
+            video_codec_name: None,
+            video_r_frame_rate: None,
+            video_width: None,
+            video_height: None,
+            bit_rate: None,
+            error: Some(e.to_string()),
+        }
     }
 }